# Conflicts:
#	Assets/_Render/Shaders.meta

+++ Assets/_Scripts/Core/Divers/GameManager.cs
@@ -5,7 +5,9 @@ using System.Collections.Generic;
 /// <summary>
 /// GameManager Description
 /// </summary>
+[RequireComponent(typeof(TutoStart))]    //tuto !!
 [RequireComponent(typeof(ScoreManager))]    //le scoreManager doit être accroché à l'objet
+[RequireComponent(typeof(ItemManager))]    //item manager
 public class GameManager : MonoBehaviour
 {
     #region Attributes
@@ -14,12 +16,28 @@ public class GameManager : MonoBehaviour
     private List<SpawnPlayer> spawnPlayer;
     public List<SpawnPlayer> SpawnPlayer { get { return spawnPlayer; } }
 
-    [FoldoutGroup("Object In World"), Tooltip("Platform mouvante"), SerializeField]
+    [FoldoutGroup("Object In World"), Tooltip("Entitées créées dynamiquement mouvantes"), SerializeField]
     private Transform objectDynamiclyCreated;
     public Transform ObjectDynamiclyCreated { get { return objectDynamiclyCreated; } }
 
+
+    [FoldoutGroup("Object In World"), Tooltip("movingPlatform"), SerializeField]
+    private MovePlatform movingPlatform;
+
+    [FoldoutGroup("Object In World"), Tooltip("panel Canvas des joueurs (in game)"), SerializeField]
+    private GameObject panelCanvasInGame;
+
+
+    [FoldoutGroup("Object In World"), Tooltip("Entitées créées dynamiquement fixes"), SerializeField]
+    private Transform objectNotMovingDynamiclyCreated;
+    public Transform ObjectNotMovingDynamiclyCreated { get { return objectNotMovingDynamiclyCreated; } }
+
+
     [FoldoutGroup("Debug"), Tooltip("Mouvement du joueur"), SerializeField]
     private GameObject prefabsPlayer;
+
+    
+    //public MovePlatform MovingPlatform { get { return movingPlatform; } }
 
     [FoldoutGroup("Debug"), Tooltip("optimisation fps"), SerializeField]
 	private FrequencyTimer updateTimer;
@@ -33,6 +51,9 @@ public class GameManager : MonoBehaviour
 
     private ItemManager itemManager;  //reference de l'itemManager;
     public ItemManager ItemManager { get { return itemManager; } }
+
+    private TutoStart tutoStart;  //reference de l'itemManager;
+    public TutoStart TutoStart { get { return tutoStart; } }
 
     //singleton
     private static GameManager instance;
@@ -57,10 +78,17 @@ public class GameManager : MonoBehaviour
     {
         SetSingleton();
         playerConnect = PlayerConnected.GetSingleton;
+        tutoStart = GetComponent<TutoStart>();
         scoreManager = GetComponent<ScoreManager>();
         itemManager = GetComponent<ItemManager>();
         if (!objectDynamiclyCreated)
             Debug.LogError("error");
+    }
+
+    private void Start()
+    {
+        ActiveGame(false);  //desactive tout au start au cas ou
+        tutoStart.ActiveTuto(true); //active les tutos
     }
 
     #endregion
@@ -73,7 +101,7 @@ public class GameManager : MonoBehaviour
     /// </summary>
     public void updateJoypadDisconnect(int id, bool connected)
     {
-        if (!connected)
+        if (!connected && spawnPlayer[id] && spawnPlayer[id].PlayerController)
             spawnPlayer[id].PlayerController.Kill();
     }
 
@@ -89,6 +117,28 @@ public class GameManager : MonoBehaviour
             SceneChangeManager.GetSingleton.Quit();
         }
     }
+
+    [FoldoutGroup("Debug")]
+    [Button("ActiveGame")]
+    public void ActiveGame() { ActiveGame(true); }
+
+    public void ActiveGame(bool active)
+    {
+        panelCanvasInGame.SetActive(active);                //active le canvas des scores
+        movingPlatform.IsScrollingAcrtive = active;         //active la platforme mouvante
+        if (!active)
+        {
+            for (int i = 0; i < spawnPlayer.Count; i++)
+            {
+                spawnPlayer[i].gameObject.SetActive(active);    //active les spawn des players
+            }
+        }
+        else
+        {
+            tutoStart.ActiveTuto(false);
+            tutoStart.enabled = false;
+        }
+    }
     #endregion
 
     #region Unity ending functions
@@ -100,7 +150,7 @@ public class GameManager : MonoBehaviour
 
         }
 
-        Quit();
+        Quit(); //input quitter
         
     }
